+++ tests/testthat/test-build-process.r
@@ -59,7 +59,7 @@ test_that("source builds return correct filenames", {
   dir.create(tmp <- tempfile())
   on.exit(unlink(tmp, recursive = TRUE), add = TRUE)
 
-  pr <- pkgbuild_process$new("testWithSrc", dest_path = tmp)
+  pr <- pkgbuild_process$new("testWithSrc", dest_path = tmp, register_routines = FALSE)
   pr$wait(60000)
   if (pr$is_alive()) {
     pr$kill()
@@ -74,7 +74,7 @@ test_that("source builds return correct filenames", {
 test_that("build package with src requires compiler", {
   without_compiler({
     expect_error({
-      pr <- pkgbuild_process$new("testWithSrc", dest_path = tempdir())
+      pr <- pkgbuild_process$new("testWithSrc", dest_path = tempdir(), register_routines = FALSE)
       pr$kill()
     }, "Could not find tools")
   })